move common logic for getRawData to shared file
