+++ backend/server/routes/reports/trainingAndQualifications/generateTrainingAndQualificationsReport.js
@@ -24,6 +24,7 @@ const {
 const { generateExpiredTrainingTab } = require('./expiredTrainingTab');
 const { generateTrainingRecordDetailsTab } = require('./trainingRecordDetailsTab');
 const { generateQualificationRecordDetailsTab } = require('./qualificationRecordDetailsTab');
+const { getRawDataForTrainingAndQualificationsReport } = require('./getRawData');
 
 const generateTrainingAndQualificationsReport = async (req, res) => {
   try {
@@ -76,28 +77,6 @@ const generateTrainingAndQualificationsReport = async (req, res) => {
     console.error(error);
     return res.status(500).send();
   }
-};
-
-const getRawDataForTrainingAndQualificationsReport = async (establishmentId, isParent = false) => {
-  const rawEstablishmentTrainingBreakdowns = await models.establishment.workersAndTraining(
-    establishmentId,
-    true,
-    isParent,
-  );
-  const rawEstablishmentWithTrainingRecords = await models.establishment.getEstablishmentTrainingRecords(
-    establishmentId,
-    isParent,
-  );
-  const rawEstablishmentCareCertificateStatus = await models.establishment.getWorkersWithCareCertificateStatus(
-    establishmentId,
-    isParent,
-  );
-
-  return {
-    rawEstablishmentTrainingBreakdowns,
-    rawEstablishmentWithTrainingRecords,
-    rawEstablishmentCareCertificateStatus,
-  };
 };
 
 router
